Refactor pagination data to be more useful and less for show
