+++ src/components/game/gameListDisplayer.tsx
@@ -5,13 +5,13 @@ import {PageControl} from "../pagination/pageControl";
 
 export function GameListDisplayer() {
     const [page, setPage] = useState(1)
-    const {games, loading, paginationData: {TotalPages}} = useGamesFetch(page);
+    const {games, loading, paginationData} = useGamesFetch(page);
     
     if (loading) {
         return <div>Loading...</div>
     }
     return <>
         <GameList games={games}/>
-        <PageControl Page={page} TotalPages={TotalPages} SetPage={setPage}/>
+        <PageControl Page={page} TotalPages={paginationData?.TotalPages ?? 0} SetPage={setPage}/>
     </>;
 }
